Move ClickableBuilder to CobbleSword
